Add auth: add firestore

+++ lib/data/entities/product.dart
@@ -25,13 +25,13 @@ class ProductsData {
 class ProductData {
   String name;
   double calories;
-  int servingSizeG;
+  double servingSizeG;
   double fatTotalG;
-  int fatSaturatedG;
+  double fatSaturatedG;
   double proteinG;
-  int sodiumMg;
-  int potassiumMg;
-  int cholesterolMg;
+  double sodiumMg;
+  double potassiumMg;
+  double cholesterolMg;
   double carbohydratesTotalG;
   double fiberG;
   double sugarG;
